pre-refactoring for 4335
