+++ hoot-core/src/main/cpp/hoot/core/visitors/geometrymodifiers/WayToPolyGeoModifierAction.cpp
@@ -28,7 +28,7 @@
 #include "WayToPolyGeoModifierAction.h"
 
 // Hoot
-#include <hoot/core/util/CoordinateExt.h>
+#include <hoot/core/geometry/CoordinateExt.h>
 #include <hoot/core/util/Factory.h>
 
 #include <math.h>
